+++ src/features/visualizer/import/implementations/json.tsx
@@ -161,6 +161,7 @@ export const ImportJSON: ImportOption = {
     values,
     controller,
   }: {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     values: Record<string, any>;
     controller: VisualizerStore["controller"];
   }) => {
@@ -193,7 +194,9 @@ export const ImportJSON: ImportOption = {
       await controller.db.saveDatabase();
       return result;
     } catch (err) {
-      await controller.db.deleteDatabase(databaseName);
+      if (databaseCreated) {
+        await controller.db.deleteDatabase(databaseName);
+      }
       throw err;
     }
   },
